Divide Move into position and line
