+++ python/linepost/position.py
@@ -1,4 +1,4 @@
-import linepost.move as lpmove
+# doc
 
 
 class Position:
@@ -9,7 +9,7 @@ class Position:
     def make_move(self, move, evaluation=None, remarks=None):
         next_board = self.board.copy()
         next_board.push_san(move)
-        self.moves.append(lpmove.Move(self.board, next_board,
+        self.moves.append(Move(self.board, next_board,
                                       move, evaluation, remarks))
         return Position(next_board)
 
@@ -18,3 +18,19 @@ class Position:
             raise ValueError('Cannot merge differing positions')
 
         self.moves.extend(other.moves)
+
+
+class Move:
+    """Move represents the transition from one position to another.
+    """
+
+    def __init__(self, from_position, to_position, label,
+                 evaluation=None, remarks=None):
+        self.from_position = from_position
+        self.to_position = to_position
+        self.label = label
+        self.evaluation = evaluation
+        if remarks:
+            self.remarks = remarks
+        else:
+            self.remarks = []
